Some little changes

+++ ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: mpauw <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/08 08:28:30 by mpauw             #+#    #+#             */
-/*   Updated: 2017/11/10 11:55:22 by mpauw            ###   ########.fr       */
+/*   Updated: 2018/03/07 17:12:50 by mpauw            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ static int	recur(int nb, char **c, int i)
 	return (0);
 }
 
-char		*ft_itoa(int n)
+char		*ft_itoa(long long int n)
 {
 	int		neg;
 	int		length;
